Fix casts and types and fix unintented change to Address constructor.

+++ src/hotspot/cpu/aarch64/assembler_aarch64.hpp
@@ -404,7 +404,7 @@ class Address {
     : _mode(no_mode) { }
   Address(Register r)
     : _base(r), _index(noreg), _offset(0), _mode(base_plus_offset), _target(0) { }
-  Address(Register r, int64_t o)
+  Address(Register r, int o)
     : _base(r), _index(noreg), _offset(o), _mode(base_plus_offset), _target(0) { }
   Address(Register r, ByteSize disp)
     : Address(r, in_bytes(disp)) { }
